fix(cargo-tangle): fix blueprint create subcommand after merge
